+++ src/gb.cpp
@@ -20,7 +20,7 @@ namespace gameboy
 
         do
         {
-            int cycles = cpu.cycle() * 4;
+            uint8_t cycles = cpu.cycle() * 4;
             timer.cycle(cycles);
             ppu.cycle(cycles);
         } while (!updatePlatform(lastCycleTime, ppu, input));
